fix: moving analytics processors to shared and fixing them, fixing other linting errors toos
