+++ packages/common/tsconfig.json
@@ -15,7 +15,7 @@
   },
   "include": [
     "**/*"
-  ],
+, "../shared/processors/drive-analytics/index.ts", "../shared/processors/drive-analytics/types.ts"  ],
   "exclude": [
     "dist",
     "node_modules",
