+++ packages/registry/cdn-cache/@powerhousedao/vetra/6.0.0-dev.59/package.json
@@ -1,0 +1,114 @@
+{
+  "name": "@powerhousedao/vetra",
+  "version": "6.0.0-dev.59",
+  "license": "AGPL-3.0-only",
+  "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/powerhouse-inc/powerhouse"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    },
+    "./document-models": {
+      "import": "./dist/document-models/index.js",
+      "types": "./dist/document-models/index.d.ts"
+    },
+    "./document-models/*": {
+      "types": "./dist/document-models/*/index.d.ts",
+      "default": "./dist/document-models/*/index.js"
+    },
+    "./editors/*": {
+      "types": "./dist/editors/*/index.d.ts",
+      "default": "./dist/editors/*/index.js"
+    },
+    "./editors": {
+      "import": "./dist/editors/index.js",
+      "types": "./dist/editors/index.d.ts"
+    },
+    "./subgraphs": {
+      "import": "./dist/subgraphs/index.js",
+      "types": "./dist/subgraphs/index.d.ts"
+    },
+    "./processors": {
+      "import": "./dist/processors/index.js",
+      "types": "./dist/processors/index.d.ts"
+    },
+    "./manifest": {
+      "import": "./dist/manifest.js",
+      "types": "./dist/manifest.d.ts"
+    },
+    "./vetra-drive-app": {
+      "import": "./dist/editors/vetra-drive-app/index.js",
+      "types": "./dist/editors/vetra-drive-app/index.d.ts"
+    },
+    "./style.css": "./dist/style.css"
+  },
+  "sideEffects": false,
+  "scripts": {
+    "tsc": "tsc",
+    "lint": "eslint",
+    "build:css": "pnpm exec tailwindcss -i ./style.css -o ./dist/style.css",
+    "test": "vitest --run",
+    "test:watch": "vitest",
+    "test:document-models": "vitest run document-models",
+    "test:editors": "vitest run editors",
+    "test:processors": "vitest run processors",
+    "generate": "ph-cli generate",
+    "connect": "vite",
+    "reactor": "ph-cli reactor",
+    "service": "ph-cli service",
+    "migrate": "ph-cli migrate",
+    "service-startup": "bash ./node_modules/@powerhousedao/ph-cli/dist/scripts/service-startup.sh",
+    "service-unstartup": "bash ./node_modules/@powerhousedao/ph-cli/dist/scripts/service-unstartup.sh"
+  },
+  "dependencies": {
+    "@powerhousedao/builder-tools": "workspace:*",
+    "@powerhousedao/common": "workspace:*",
+    "@powerhousedao/config": "workspace:*",
+    "@powerhousedao/design-system": "workspace:*",
+    "@powerhousedao/reactor": "workspace:*",
+    "change-case": "catalog:",
+    "colorette": "catalog:",
+    "document-model": "workspace:*",
+    "graphql": "catalog:",
+    "kysely": "catalog:",
+    "@powerhousedao/shared": "workspace:*",
+    "zod": "catalog:",
+    "graphql-tag": "catalog:"
+  },
+  "devDependencies": {
+    "tailwindcss": "catalog:",
+    "@powerhousedao/codegen": "workspace:*",
+    "@powerhousedao/ph-cli": "workspace:*",
+    "@powerhousedao/reactor-api": "workspace:*",
+    "@powerhousedao/reactor-browser": "workspace:*",
+    "@tailwindcss/cli": "catalog:",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/bun": "^1.3.7",
+    "@types/node": "catalog:",
+    "@types/react": "catalog:",
+    "@types/react-dom": "catalog:",
+    "@vitejs/plugin-react": "catalog:",
+    "document-drive": "workspace:*",
+    "react": "catalog:",
+    "react-dom": "catalog:",
+    "vite": "catalog:",
+    "vitest": "catalog:"
+  },
+  "peerDependencies": {
+    "react": ">=19.0.0",
+    "react-dom": ">=19.0.0"
+  },
+  "description": ""
+}
