+++ packages/registry/storage/.verdaccio-db.json
@@ -1,0 +1,1 @@
+{"list":["@powerhousedao/vetra"],"secret":"FyORTPVvgMdVs3WED6k9q3OaAJoyeQwL"}
