+++ packages/registry/storage/@powerhousedao/vetra/package.json
@@ -1,0 +1,154 @@
+{
+	"name": "@powerhousedao/vetra",
+	"versions": {
+		"6.0.0-dev.59": {
+			"name": "@powerhousedao/vetra",
+			"version": "6.0.0-dev.59",
+			"license": "AGPL-3.0-only",
+			"type": "module",
+			"repository": {
+				"type": "git",
+				"url": "git+https://github.com/powerhouse-inc/powerhouse.git"
+			},
+			"publishConfig": {
+				"access": "public"
+			},
+			"files": [
+				"dist"
+			],
+			"exports": {
+				".": {
+					"import": "./dist/index.js",
+					"types": "./dist/index.d.ts"
+				},
+				"./document-models": {
+					"import": "./dist/document-models/index.js",
+					"types": "./dist/document-models/index.d.ts"
+				},
+				"./document-models/*": {
+					"types": "./dist/document-models/*/index.d.ts",
+					"default": "./dist/document-models/*/index.js"
+				},
+				"./editors/*": {
+					"types": "./dist/editors/*/index.d.ts",
+					"default": "./dist/editors/*/index.js"
+				},
+				"./editors": {
+					"import": "./dist/editors/index.js",
+					"types": "./dist/editors/index.d.ts"
+				},
+				"./subgraphs": {
+					"import": "./dist/subgraphs/index.js",
+					"types": "./dist/subgraphs/index.d.ts"
+				},
+				"./processors": {
+					"import": "./dist/processors/index.js",
+					"types": "./dist/processors/index.d.ts"
+				},
+				"./manifest": {
+					"import": "./dist/manifest.js",
+					"types": "./dist/manifest.d.ts"
+				},
+				"./vetra-drive-app": {
+					"import": "./dist/editors/vetra-drive-app/index.js",
+					"types": "./dist/editors/vetra-drive-app/index.d.ts"
+				},
+				"./style.css": "./dist/style.css"
+			},
+			"sideEffects": false,
+			"scripts": {
+				"tsc": "tsc",
+				"lint": "eslint",
+				"build:css": "pnpm exec tailwindcss -i ./style.css -o ./dist/style.css",
+				"test": "vitest --run",
+				"test:watch": "vitest",
+				"test:document-models": "vitest run document-models",
+				"test:editors": "vitest run editors",
+				"test:processors": "vitest run processors",
+				"generate": "ph-cli generate",
+				"connect": "vite",
+				"reactor": "ph-cli reactor",
+				"service": "ph-cli service",
+				"migrate": "ph-cli migrate",
+				"service-startup": "bash ./node_modules/@powerhousedao/ph-cli/dist/scripts/service-startup.sh",
+				"service-unstartup": "bash ./node_modules/@powerhousedao/ph-cli/dist/scripts/service-unstartup.sh"
+			},
+			"dependencies": {
+				"@powerhousedao/builder-tools": "workspace:*",
+				"@powerhousedao/common": "workspace:*",
+				"@powerhousedao/config": "workspace:*",
+				"@powerhousedao/design-system": "workspace:*",
+				"@powerhousedao/reactor": "workspace:*",
+				"change-case": "catalog:",
+				"colorette": "catalog:",
+				"document-model": "workspace:*",
+				"graphql": "catalog:",
+				"kysely": "catalog:",
+				"@powerhousedao/shared": "workspace:*",
+				"zod": "catalog:",
+				"graphql-tag": "catalog:"
+			},
+			"devDependencies": {
+				"tailwindcss": "catalog:",
+				"@powerhousedao/codegen": "workspace:*",
+				"@powerhousedao/ph-cli": "workspace:*",
+				"@powerhousedao/reactor-api": "workspace:*",
+				"@powerhousedao/reactor-browser": "workspace:*",
+				"@tailwindcss/cli": "catalog:",
+				"@testing-library/jest-dom": "^6.9.1",
+				"@testing-library/react": "^16.3.0",
+				"@testing-library/user-event": "^14.6.1",
+				"@types/bun": "^1.3.7",
+				"@types/node": "catalog:",
+				"@types/react": "catalog:",
+				"@types/react-dom": "catalog:",
+				"@vitejs/plugin-react": "catalog:",
+				"document-drive": "workspace:*",
+				"react": "catalog:",
+				"react-dom": "catalog:",
+				"vite": "catalog:",
+				"vitest": "catalog:"
+			},
+			"peerDependencies": {
+				"react": ">=19.0.0",
+				"react-dom": ">=19.0.0"
+			},
+			"description": "",
+			"gitHead": "9119223b044c6834aaff27812bec582563c2153f",
+			"_id": "@powerhousedao/vetra@6.0.0-dev.59",
+			"bugs": {
+				"url": "https://github.com/powerhouse-inc/powerhouse/issues"
+			},
+			"homepage": "https://github.com/powerhouse-inc/powerhouse#readme",
+			"_nodeVersion": "24.13.0",
+			"_npmVersion": "11.6.2",
+			"dist": {
+				"integrity": "sha512-jXLPvQGZnke2X1cMsJDsgW9k8X2/Upg+HcxBOaxL92OmVgW6amXrNKL5aaNOH/JQP7nulvcGundajGOFuRa6Zg==",
+				"shasum": "e80da5d6b67cabab444c7a0296d6292db281ad63",
+				"tarball": "http://localhost:4873/@powerhousedao/vetra/-/@powerhousedao/vetra-6.0.0-dev.59.tgz"
+			},
+			"contributors": []
+		}
+	},
+	"time": {
+		"modified": "2026-02-26T16:37:24.177Z",
+		"created": "2026-02-26T16:37:24.177Z",
+		"6.0.0-dev.59": "2026-02-26T16:37:24.177Z"
+	},
+	"users": {},
+	"dist-tags": {
+		"latest": "6.0.0-dev.59",
+		"dev": "6.0.0-dev.59"
+	},
+	"_uplinks": {},
+	"_distfiles": {},
+	"_attachments": {
+		"vetra-6.0.0-dev.59.tgz": {
+			"shasum": "e80da5d6b67cabab444c7a0296d6292db281ad63",
+			"version": "6.0.0-dev.59"
+		}
+	},
+	"_rev": "3-c4e6d819719b09ca",
+	"_id": "@powerhousedao/vetra",
+	"readme": "ERROR: No README data found!"
+}
