+++ packages/registry/storage/htpasswd
@@ -1,0 +1,1 @@
+testuser:$2a$10$.Tpmetdbz2OiUKvfeRM1Sec1OoTW3SJ7MeEO2FE/wsK9CrA47/KEy:autocreated 2026-02-26T16:37:16.190Z
