+++ packages/shared/processors/drive-analytics/index.ts
@@ -1,12 +1,12 @@
 import type { IAnalyticsStore } from "@powerhousedao/analytics-engine-core";
-import type { ProcessorRecordLegacy } from "document-drive";
-import type { PHDocumentHeader } from "document-model";
+import type { PHDocumentHeader } from "../../document-model/core/documents.js";
+import type { ProcessorRecord } from "../types.js";
 import { DocumentAnalyticsProcessor } from "./document-processor.js";
 import { DriveAnalyticsProcessor } from "./drive-processor.js";
 
 export const processorFactory =
   (module: { analyticsStore: IAnalyticsStore }) =>
-  (driveHeader: PHDocumentHeader): ProcessorRecordLegacy[] => {
+  (driveHeader: PHDocumentHeader): ProcessorRecord[] => {
     return [
       {
         processor: new DriveAnalyticsProcessor(module.analyticsStore),
