+++ tsconfig.json
@@ -74,6 +74,9 @@
       "path": "test/e2e-utils"
     },
     {
+      "path": "test/switchboard"
+    },
+    {
       "path": "test/test-client"
     },
     {
